requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
PyYAML==6.0.2